Sample parameters from a model

sample_posterior() draws samples from the posterior distribution of the specified model given a data set with visit data.

sample_prior() draws samples from the prior distribution of the specified model object.

Usage

sample_posterior(
  model,
  data,
  now = NULL,
  nsim = 2000L,
  seed = NULL,
  warmup = 500L,
  nuts_control = list(),
  acceptable_divergent_transition_fraction = 0.1,
  ...
)

sample_prior(
  model,
  nsim = 2000L,
  seed = NULL,
  warmup = 500L,
  nuts_control = list(),
  ...
)

Arguments

model

an object of class srpmodel containing prior information

data

a data frame with variables subject_id<chr> (subject identifier), group_id<chr> (group identifier), t<dbl> (time of visit, relative to first visit in study), state<chr> (state recorded at visit). Allowed states are "stable", "response", "progression" (or death), and "EOF" (end of follow-up). The EOF state marks the end of an individual's follow-up before the absorbing state "progression".

now

numeric, time from first visit in data if different form last recorded visit

nsim

integer, number of samples to draw

seed

integer, fixed random seed; NULL for no fixed seed

warmup

integer, number of warm-up samples for the MCMC sampler before retaining samples; see warmup parameter in rstan::stan().

nuts_control

list, parameters for NUTS algorithm see control argument inrstan::stan()

acceptable_divergent_transition_fraction,

numeric between 0 and 1 giving the acceptable fraction of divergent transitions before throwing an error

...

further arguments passed to method implementations

Value

A rstan::stanfit object with posterior samples.

See also

Examples

mdl <- create_srpmodel(A = define_srp_prior())
tbl <- tibble::tibble(
  subject_id = c("A1", "A1"),
  group_id = c("A", "A"),
  t = c(0, 1.5),
  state = c("stable", "response")
)
sample_posterior(mdl, tbl, seed = 42L)
#> Inference for Stan model: srp_model_simple.
#> 1 chains, each with iter=2500; warmup=500; thin=1; 
#> post-warmup draws per chain=2000, total post-warmup draws=2000.
#> 
#>                mean se_mean    sd   2.5%   25%   50%   75% 97.5% n_eff Rhat
#> p[1]           0.63    0.00  0.21   0.19  0.49  0.65  0.79  0.96  3312    1
#> median_t[1,1]  3.29    0.13  4.85   0.51  1.18  1.94  3.64 14.41  1293    1
#> median_t[1,2]  8.42    0.22 10.40   0.69  2.48  4.85 10.12 37.04  2158    1
#> median_t[1,3] 10.88    0.35 14.26   0.99  3.25  6.42 12.48 51.66  1661    1
#> shape[1,1]     1.53    0.01  0.46   0.82  1.21  1.46  1.80  2.56  2486    1
#> shape[1,2]     1.57    0.01  0.47   0.84  1.23  1.50  1.83  2.64  2450    1
#> shape[1,3]     1.58    0.01  0.51   0.81  1.21  1.50  1.88  2.74  2805    1
#> scale[1,1]     4.39    0.20  6.97   0.65  1.49  2.54  4.78 19.43  1260    1
#> scale[1,2]    10.87    0.28 13.21   0.89  3.17  6.30 13.35 46.82  2189    1
#> scale[1,3]    14.15    0.46 18.62   1.27  4.15  8.26 16.54 70.60  1646    1
#> lp__          -8.31    0.07  1.98 -13.18 -9.32 -7.95 -6.88 -5.58   740    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Dec  8 19:20:00 2022.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

sample_prior(mdl, seed = 42L)
#> Inference for Stan model: srp_model_simple.
#> 1 chains, each with iter=2500; warmup=500; thin=1; 
#> post-warmup draws per chain=2000, total post-warmup draws=2000.
#> 
#>                mean se_mean    sd   2.5%   25%   50%   75% 97.5% n_eff Rhat
#> p[1]           0.50    0.00  0.25   0.06  0.30  0.50  0.69  0.93  2820    1
#> median_t[1,1] 10.69    0.31 13.42   0.69  2.93  5.88 12.82 49.80  1861    1
#> median_t[1,2]  7.92    0.21  9.24   0.74  2.59  5.01  9.56 34.01  2010    1
#> median_t[1,3] 10.81    0.39 14.92   0.73  2.95  5.87 12.22 53.42  1496    1
#> shape[1,1]     1.57    0.01  0.49   0.85  1.20  1.48  1.85  2.72  2775    1
#> shape[1,2]     1.59    0.01  0.52   0.82  1.20  1.50  1.87  2.83  3220    1
#> shape[1,3]     1.57    0.01  0.52   0.81  1.21  1.48  1.85  2.76  2841    1
#> scale[1,1]    13.78    0.40 17.29   0.89  3.78  7.64 16.42 64.25  1863    1
#> scale[1,2]    10.26    0.27 11.96   0.94  3.36  6.43 12.45 45.06  2018    1
#> scale[1,3]    14.11    0.52 19.79   0.90  3.85  7.61 15.81 71.18  1424    1
#> lp__          -6.08    0.07  1.99 -10.68 -7.21 -5.68 -4.59 -3.29   905    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Dec  8 19:20:00 2022.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).